+++ internal/hook/shortsocket_windows_test.go
@@ -1,6 +1,0 @@
-//go:build windows
-
-package hook
-
-// shortSocketBase is unused on Windows: the socket-marker case skips there.
-const shortSocketBase = ""